login.php done, misc functions in miscfunctions
Good idea to include miscfunctions on every page, that way everyone can
have access to functions one of us comes up with. I've written
valid($string), which is a function to check for valid inputs, and
loggedIn(), which returns a boolean for as to whether or not a user is
logged in.